+++ springboot/src/main/java/hello/controller/EventController.java
@@ -1,5 +1,6 @@
 package hello.controller;
 
+import hello.OkResponse;
 import hello.entity.Event;
 import hello.repository.EventRepository;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -37,19 +38,5 @@ class GetEventsListResponse {
 
 	public List<Event> getEventsList() {
 		return eventsList;
-	}
-}
-
-class OkResponse {
-	private Object data;
-
-	public OkResponse(Object data) {
-		this.data = data;
-	}
-
-	public Integer getStatus() { return 200; }
-
-	public Object getData() {
-		return data;
 	}
 }
